Move state machine to one file
